backend/requirements.txt
fastapi
uvicorn
google-generativeai
cohere
pinecone
asyncpg
python-dotenv
pydantic
pydantic-settings
langchain-text-splitters
tiktoken
python-multipart
PyJWT
passlib[bcrypt]
